refactor(login and signup): Refactor booth components into pages.
