+++ src/app/pages/login/login.component.scss
@@ -1,4 +1,3 @@
-/* src/app/auth/login/login.component.scss */
 ion-content {
   --background: var(--ion-background-color);
 }
@@ -8,6 +7,7 @@ ion-content {
   margin: 0 auto;
   padding: 20px 0;
 }
+
 .header-section {
   margin-bottom: 30px;
