+++ src/app/pages/login/login.component.ts
@@ -2,8 +2,8 @@
 import { Component } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { IonicModule } from '@ionic/angular';
-import { AuthService } from '../../../Services/auth.service';
-import { FormComponent, LoginFormData } from '../form/login/form.component';
+import { AuthService } from '../../Services/auth.service';
+import { FormComponent, LoginFormData } from '../../components/Auth/form/login/login-form.component';
 
 @Component({
   selector: 'app-login',
